fix(alsa): properly set buffer size in ALSA configuration to get the correct requested latency
